refactoring par 2
