+++ src/lib/api/index.ts
@@ -1,0 +1,7 @@
+// API utilities exports
+export * from "./auth";
+export * from "./response";
+export * from "./request";
+export * from "./database";
+export * from "./error-transaction";
+export * from "./logging-transaction";
